payment

+++ RestaurantRaterBooking/Areas/RestaurantUser/Controllers/RestaurantsController.cs
@@ -174,7 +174,10 @@ namespace RestaurantRaterBooking.Areas.RestaurantUser.Controllers
 
 				foreach (var existingImage in imagesToDelete)
 				{
-					DeleteFile(existingImage.ImagePath);
+					if (!existingImage.ImagePath.Contains("NoImage"))
+					{
+						DeleteFile(existingImage.ImagePath);
+					}
 					_context.Image.Remove(existingImage);
 				}
